Update `stacks.boost` to return abs(BL) and other fixes.
BL seems to avoid numerical errors that BR has, so for symmetric stacks, calculating BL seems like a reasonable way to avoid these errors.

The real IoRs for which the stack is resonant (`n1_real0` and `n2_real0`), which determine the stack layer dimensions, can now be set by hand in `stacks.hw_boost`. `n2_real0` is now set to 1 in the examples in `n_calcs`. I think this is better than setting it to `n2(omega0)`, which is a bit arbitrary since the resonance may not be centered exactly on `omega0`. This keeps the stack resonant when the gas IoR is near a typical value of 1 (and `omega` is about `omega0`).

(Also fixed a unit issue with `omega_norm2`, which is now deleted. Now the effect of the molecular resonance is visible even using the big scale for omega (1st graph).)