Create plots for the jan 5 (ish) update.
Also correct a factor of 1e-6 to 1e-3 in `CO_n`, update `stacks._hw_boost` to work for a single middle layer, and add the `if __name__ == '__main__':` idiom in `n_calcs`.

IMPORTANT: The DM-induced electric field, `A`, does not yet correctly account for the changes in the refractive index of the gas layer.